requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
seaborn
Pillow
